Debug gpiorelay
When two GPIORelay's are created successively using the same name, the first one's __del__
is called as the reference to that first object is now gone but the second is not
initialized properly.  Is this just Python optimizing, or am I misunderstanding what's going
on?

Code:
a = GPIORelay(default_input=1)
# create a new one with the same name
a = GPIORelay(default_input=2)
a.select_input(1)

Expected:
'a' is created, set to input 1, destroyed, created again, set to input 2, then set back to 1

Actual:
'a' is created, set to input 1, destroyed, created again but then an exception is raised
while setting the input to 2 as GPIO.setmode never executes during the second __init__.
Maybe there is no second call to __init__ and it's just reusing the same object?  This
exception causes the program to terminate, at which point the second 'a' is destroyed,
calling GPIO.cleanup with an already clean slate.  This prints a warning telling you to
try calling GPIO.cleanup only at program exit.

This does not really break anything major but I want to investigate further and figure this out.