nonblocking socket
Working on socket communication to try to make it work just as well as serial.  The issue
is that any message that contains '\r\n' is only read up until that sequence, so
input_status only sees the first part of the reply and thinks there is only a single output
on a two-output switcher.  I have added a small (0.05s) delay to recv() inside the select
while loop.  I think that might fix it.. Gotta test.