implement custom input support
Since not all NEC projectors have the same inputs or use the same code for every input,
an 'inputs' parameter has been added to the constructor.  If present, this must be a dict
of new input values.  This dict will be merged with the existing values in NEC.Input at
__init__ and the result assigned to a new Enum meta called 'MyInputs'.  self.my_inputs is
a reference to this.  If no input dict is given, self.my_inputs is a reference to NEC.Input.
All input selection, querying, and validation therefore goes through self.my_inputs now.