more KramerVP734 debugging
with the changes I've made to the socket operations, the value of _input_status returned
by the call to select_input() reflects the previous input status from before the call.
I think it's seeing old output, so I'm testing resetting the input buffer prior to the call.

above didn't work. try adding some sleep time back to select_input before it reads the
response.

(It is really annoying having to develop on one machine and test on another.  It means I
have to git push every little change so I can test it)