more KramerVP734 debugging
with the changes I've made to the socket operations, the value of _input_status returned
by the call to select_input() reflects the previous input status from before the call.
I think it's seeing old output, so I'm testing resetting the input buffer prior to the call.

above didn't work. try adding some sleep time back to select_input before it reads the
response.

Still didn't work. This is kindof a low priority fix because we're probably not going to
be using these devices over ethernet much and it works fine over rs232.  For now, the fix
may be to call self.input_status (which has been working fine) after self.select_input, instead
of returning the internal state variable _input_status.  It appears _input_status is not
getting set properly after a select_input() call, probably because read_response isn't
reading everything like it was when I was creating the socket in blocking mode.

(It is really annoying having to develop on one machine and test on another.  It means I
have to git push every little change so I can test it)