Refactoring on all drivers
1. Got rid of enums for input mappings as it was clunky and unnecessary.  How do you check
an object for membership in an enum created at runtime?  `str(type(x))=='<class Input>'` is
the only way that comes to mind.  Also, enums can't be returned from a flask view function,
so we might as well just be dealing with strings.

2. All input dicts have been normalized to use strings as keys and either strings or
ints as values.  utils.key_for_value is used to obtain the first key that is assigned to the
given value.  All custom input names appear before the defaults in the merged dicts,
meaning the input's custom name is what's returned if it's present in the mapping.

3. Reintroduced the Projector, Switcher, and TV classes as they can actually serve a purpose
now.  Debating on moving them out of the drivers package and into the blueprints as that is
where they will be created and used.
When the application reads the equipment configuration, each device should have a
mapping of input names to values.  This mapping is stored in the wrapper and then passed
into the constructor for said wrapper's driver.  Because only the config-specified names get
saved in the wrapper, the interface need not clutter the screen with every input listed in
the driver, only those our device is actually using.  Also, each device wrapper can be given
a default input that is selected whenever the system is first turned on or reset.