add key checking on select_input()
Because I guess by default a KeyError just contains the bad key and no other text (why?),
attempting to navigate to '/input/an_input_that_does_not_exist' was flashing just the text
'an_input_that_does_not_exist' as an error.  And because it doesn't actually contain the
string 'error', it was being flashed in green like everything was way cool.  This is fixed.
Errors need to be red.  And explain something.