make most of the buttons work
Yay! It actually does stuff now.  Implemented all projector controls.  Projector button opens
the projector index/home/status page.  On/reset powers on the projector (and selects the
default input if it is set), power off kills it, and the input buttons do their thing.
Been testing with the PJLink driver and I'm pretty happy with the results so far.

Next, I need to implement the home buttons and the system reset/power off.