clumsy hack for system reset
NEC driver it seems can switch inputs ~2 seconds after power on.  So system_reset waits 2
seconds after power on, then switches input.  PJLink cannot do that... the device has to
completely finish startup before taking other commands.  As a clumsy hack, we check for a
CommandFailureError and tell the user to wait until the projector is done warming up.
Short of implementing multi-threading I really can't think of any way to queue up the
projector to switch inputs whenever it's ready without blocking for a solid minute.
We may need to just scrap the idea of powering on & selecting a device all at once and stick
with separate actions.