a bit of refactoring
Was proving difficult to have buttons on the home page perform both a power on and input
selection function, particularly when different drivers (NEC vs PJLink) require a different
amount of time before the device will respond to subsequent commands.  For now, I have
attempted to clean up the PJ error / warning messages a little and better indicate to the
user that if a display is powered off, they need to power it on first.

Interface looks a little cleaner and I've implemented some bootstrap modal dialogs for
confirming power on/power off as these are costly mistakes to make time-wise and money-wise
when your lamp eventually dies from overstriking.